requirements.txt
aniso8601==8.0.0
appdirs==1.4.3
attrs==20.3.0
backcall==0.2.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
flask-restplus==0.13.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
parso==0.7.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
progress==1.5
prompt-toolkit==3.0.8
PTable==0.9.2
ptyprocess==0.6.0
Pygments==2.7.2
pyparsing==2.4.6
pyrsistent==0.17.3
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.4
redis==3.5.3
redisgraph==2.2.3
requests==2.22.0
retrying==1.3.3
six==1.14.0
traitlets==5.0.5
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.1
